+++ NEWS.md
@@ -56,6 +56,8 @@
 * New `compression_level` option to select the compression level
   manually. See `?parquet_options` for details. (#91).
 
+* `read_parquet()` can now read from an R connection (#71).
+
 * `read_parquet()` now reads `DECIMAL` values correctly from `INT32`
   and `INT64` columns if their `scale` is not zero.